logic-worker/requirements.txt
kafka-python
openai
requests
psycopg2-binary
python-dotenv
